Реализовано: Авторизация, валидация запросов, обработчик exception-ов

+++ src/main/java/com/news/repository/NewsRepository.java
@@ -3,6 +3,6 @@ package com.news.repository;
 import com.news.domain.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NewsRepository extends JpaRepository<News, Integer> {
+public interface NewsRepository extends JpaRepository<News, Long> {
 
 }
